Authentication middleware modified according to new token login functionality
